Move instant production into a module.
