+++ src/maps/london/settings.ts
@@ -7,11 +7,11 @@ import { map } from "./grid";
 import { LondonStarter } from "./starter";
 import { LondonAllowedActions } from "./allowed_actions";
 import { LondonPhaseEngine } from "./production";
-import { LondonMoveAction, LondonMovePhase } from "./move_good";
 import { LondonShareHelper } from "./shares";
 import { LondonUrbanizeAction } from "./urbanize";
 import { LondonPlayerHelper } from "./score";
 import { TurnLengthModule } from "../../modules/turn_length";
+import { InstantProductionModule } from "../../modules/instant_production/module";
 
 export class LondonMapSettings implements MapSettings {
   static readonly key = GameKey.LONDON;
@@ -33,8 +33,6 @@ export class LondonMapSettings implements MapSettings {
       LondonStarter,
       LondonBuilderHelper,
       LondonPhaseEngine,
-      LondonMoveAction,
-      LondonMovePhase,
       LondonShareHelper,
       LondonUrbanizeAction,
       LondonPlayerHelper,
@@ -42,6 +40,6 @@ export class LondonMapSettings implements MapSettings {
   }
 
   getModules() {
-    return [new TurnLengthModule({ add: -1 })];
+    return [new TurnLengthModule({ add: -1 }), new InstantProductionModule()];
   }
 }
